pyproject.toml
[project]
name = "semantic-object-stream-discovery"
version = "0.1.0"
description = "Unsupervised vision pipeline for object clustering and flow tracking in video streams"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Video processing
    "opencv-python>=4.8.0",
    "av>=12.0.0",  # For HLS/RTSP streams
    "yt-dlp>=2024.0.0",  # YouTube (handles protected videos)
    "streamlink>=6.0.0",  # YouTube/Twitch streams (fallback)

    # Deep learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",  # For DINOv2
    "ultralytics>=8.0.0",  # YOLOv8 for base detection

    # Clustering & ML
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.0",

    # Tracking
    "filterpy>=1.4.5",  # Kalman filters
    "scipy>=1.11.0",
    "lap>=0.4.0",  # Linear assignment for tracking

    # Analytics & Visualization
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",

    # GCP
    "google-cloud-pubsub>=2.18.0",
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-storage>=2.13.0",

    # Utils
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sosd = "src.main:main"
sosd-dashboard = "src.output.dashboard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
